Use Pool.map_async and get() result, rather than Pool.apply_async.
This will check for exceptions during the transcoding process, rather
than silently ignoring them. It also allows the user to interrupt the
current transcode with Ctrl-C.

Note: the use of Pool.map_async followed by get() with a large
timeout, rather than just Pool.map, is a workaround for a bug in
Python's implementation of Pool.join(). See the comments in the code
for details.