Clean up children of pool processes on SIGTERM.
To ensure that a terminated pool subprocess terminates its children,
make each pool subprocess a process group leader, and install a
SIGTERM handler in the pool subprocess that kills the process group.

Prior to this change, child processes of pool processes would continue
until finished, if the pool process was terminated.